Added important information and smaller fixes.
I've added important information about HaveIBeenPwned? to mention that it also scans for passwords and added that FFProfile is interactive and configurable. I also fixed up some of the grammar and added dashes to separate the website/product name and the description. I also slightly changed uBlock Origin's description to emphasize how essential it is as a content and tracker blocker.